Modified signature of recess main function to not require an array of applications.
Modified the functionality of Make::a and Make::an to accept constructor arguments.

Taking out the trash.

Recess Tools no longer responsible for adding a parts path within Recess. This is now happening in the bootstrap process with the new UI bits.

Added the __tostring method on PathFinder to return the last added path as a string.

New capabilities of an application: initialize plugins, static reference to the running application, views now uses PathFinder for view path lookup.

Plugin abstract base class with an init method to allow registration of hooks.

Use the active application's path finder.

Registering a dependency in Buffer helper.

Introducing ArrayBlock for array/map-like access to a block with nested blocks.

Changed assignment using magic call method to explicit get(), set() that take a string as first param. Avoids naming collisions with the other methods of the class, also less magic.

Changing NativeView to use Pathfinder vs. a static path.

Removing split method which is deprecated in php 5.3

Additional method to response which allows for an array of cookies to be added at once.

Conflict response no longer an error response.

Removing temporary 'each' parts.

Introducing Control, Container, and UI Parts.

Comments for PathFinder